🐛 fix: ensure float type consistency for totalAmount in API responses
- Convert totalAmount values to float in PaymentsSummary creation
- Update ProcessorSummary model to use float instead of Decimal
- Prevents type inconsistencies between storage and API responses

+++ src/domain/models.py
@@ -16,7 +16,7 @@ class PaymentResponse(BaseModel):
 
 class ProcessorSummary(BaseModel):
     totalRequests: int
-    totalAmount: Decimal
+    totalAmount: float
 
 
 class PaymentsSummary(BaseModel):
